Editing Contacts

+++ src/com/bridgelabz/AddressBookContacts.java
@@ -74,11 +74,11 @@ public class AddressBookContacts {
         this.email = email;
     }
 
-    public void setZip(int zip) {
+    public void setZip(long zip) {
         this.zip = zip;
     }
 
-    public void setPhoneNo(int phoneNo) {
+    public void setPhoneNo(long phoneNo) {
         this.phoneNo = phoneNo;
     }
